% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_builders.R
\name{add_ep}
\alias{add_ep}
\title{Add endpoint to end of URL}
\usage{
add_ep(url, ep, id = "")
}
\arguments{
\item{url}{url}

\item{ep}{which endpt}

\item{id}{optional, an id for the endpoint, defaults to ""}
}
\value{
url with endpoint added

add_ep("http://google.com", "endpoint")
add_ep("http://google.com", "endpoint", "id")
}
\description{
Add endpoint to end of URL
}
