% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_builders.R
\name{add_token}
\alias{add_token}
\title{Add personal access token to URL}
\usage{
add_token(url, token)
}
\arguments{
\item{url}{url}

\item{token}{access token}
}
\value{
url w/ access token as parameter

add_token("https://a_url", "secret_token_xxxx")
}
\description{
Add personal access token to URL
}
