% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jcb.R
\name{create_clean_jcb_data}
\alias{create_clean_jcb_data}
\title{Import myJCB CSV data}
\usage{
create_clean_jcb_data(path, encoding = "cp932")
}
\arguments{
\item{path}{path to csv file}
}
\value{
dataframe
}
\description{
Import myJCB CSV data
}
