% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ynab_functions.R
\name{execute_get_req}
\alias{execute_get_req}
\title{Execute a YNAB GET request}
\usage{
execute_get_req(endpoint, timeout = 20)
}
\arguments{
\item{entry_point}{}
}
\value{
what ever obect has been requested
}
\description{
This function executes a GET request against the YNAB api.
}
