% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_builders.R
\name{get_url}
\alias{get_url}
\title{Get full URL}
\usage{
get_url(token, ep, id = "", within = "", within_id = "")
}
\arguments{
\item{token}{access token}

\item{ep}{which endpoint?}

\item{id}{endpoint id if just one element}

\item{within}{get endpoint within one element}

\item{within_id}{id for within}
}
\value{
full url, character of length 1
}
\description{
Get full URL
}
\examples{
get_url("secretxxxx", "budgets")
get_url("secretxxxx", "budgets", 1244)
}
