% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getToken.R
\name{getToken}
\alias{getToken}
\title{A function that asks the user to select the budget of interest}
\usage{
getToken()
}
\description{
A function that asks the user to select the budget of interest
}
\examples{
auth_token <- getToken()
}
\keyword{getToken}
