% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUnsplit.R
\name{getUnsplit}
\alias{getUnsplit}
\title{Splits multiple category transactions stored in one row into individual rows}
\usage{
getUnsplit(d)
}
\arguments{
\item{d}{name of transactions data frame of interest}
}
\description{
Splits multiple category transactions stored in one row into individual rows
}
\examples{
df <- getUnsplit(df_transactions) 
}
\keyword{getUnsplit}
