% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYNAB.R
\name{getYNAB}
\alias{getYNAB}
\title{Connects to YNAB API and converts JSON data to R data frames
JSON content > text format > list format > R data frames
Caution: Must be careful with use, as to not overal YNAB's servers}
\usage{
getYNAB(YNAB.url)
}
\arguments{
\item{YNAB.url}{Character vector of the endpoint URL of interest}
}
\description{
Connects to YNAB API and converts JSON data to R data frames
JSON content > text format > list format > R data frames
Caution: Must be careful with use, as to not overal YNAB's servers
}
\examples{
df <- getYNAB("https://api.youneedabudget.com/v1/user") 
}
\keyword{getYNAB}
