% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plastic-pasmo.R
\name{parse_plastic_pasmo_card}
\alias{parse_plastic_pasmo_card}
\title{Clean up the raw data imported from IC Reader into YNAB format}
\usage{
parse_plastic_pasmo_card(.data, purchase_label = "物販")
}
\arguments{
\item{.data}{Data frame built from the output of IC Reader app}

\item{purchase_label}{(optional) generic item label}
}
\value{
Data frame containing Pasmo card transactions in YNAB format
}
\description{
Clean up the raw data imported from IC Reader into YNAB format
}
