% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jcb.R
\name{process_jcb}
\alias{process_jcb}
\title{Import myJCB CSV data}
\usage{
process_jcb(path, encoding = "cp932")
}
\arguments{
\item{path}{path to csv file}
}
\value{
dataframe
}
\description{
Import myJCB CSV data
}
