% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobile-suica.R
\name{process_mobile_suica}
\alias{process_mobile_suica}
\title{Read and cleanup Mobile SUICA data}
\usage{
process_mobile_suica(path, year)
}
\arguments{
\item{path}{path to CSV file to cleanup}

\item{year}{Year}

\item{out_file}{(optional) name of output file}
}
\description{
Read and cleanup Mobile SUICA data
}
