% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plastic-pasmo.R
\name{process_plastic_pasmo_card}
\alias{process_plastic_pasmo_card}
\title{Read and cleanup Pasmo data from IC Reader}
\usage{
process_plastic_pasmo_card(path, out_file = "pasmo_clean.csv")
}
\arguments{
\item{path}{path to CSV file to cleanup}

\item{out_file}{(optional) name of output file}
}
\description{
Read and cleanup Pasmo data from IC Reader
}
