% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeColumnprefix.R
\name{removeColumnprefix}
\alias{removeColumnprefix}
\title{Fix column names}
\usage{
removeColumnprefix(x)
}
\arguments{
\item{x}{an R object, specifically an endpoint data frame}
}
\description{
Removes the prefix "data.{endpoint}." from column names in data frames created by getYNAB()
The assumption is that the name will follow the structure "data.{x}.something"
eg.  "data.{x}.something" > "something" or
     "data.something" > "something" or
     "something" > "something"
}
\examples{
df_transactions <- getYNAB("https://api.youneedabudget.com/v1/transactions") \%>\% removeColumnprefix()
}
\keyword{removeColumnprefix}
