% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectBudget.R
\name{selectBudget}
\alias{selectBudget}
\title{Select a budget}
\arguments{
\item{param.token}{Your YNAB API personal access token}
}
\description{
A function that asks the user to select the budget of interest
}
\keyword{selectBudget}
