% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ynab_functions.R
\name{set_ynab_token}
\alias{set_ynab_token}
\title{Set your YNAB API token}
\usage{
set_ynab_token(token)
}
\arguments{
\item{token}{}
}
\value{
No return value
}
\description{
This function sets a global option ynab_token for use in
many other function calls.  It should be the first function
called as other functions will not work without a proper
token.
}
\examples{
set_ynab_token("test************************************************************")
}
