% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ynab_functions.R
\name{summary.budget_data}
\alias{summary.budget_data}
\title{Budget Data Summary}
\usage{
\method{summary}{budget_data}(bd)
}
\arguments{
\item{bd}{a budget_data object}
}
\value{

}
\description{
Prints a summary of the budget_data object including key budget, account,
payee, category, and transaction summary information.
}
