% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ynab_functions.R
\name{ynab_get_budget}
\alias{ynab_get_budget}
\title{Get budget data}
\usage{
ynab_get_budget(budget, remove_deleted = TRUE, remove_closed = TRUE)
}
\arguments{
\item{budget}{the name or id of an available budget. It can be determined by
using the ynab_list_budgets() command.}
}
\value{
a "budget_data" object representing a list of the various data returned
by the YNAB API, with only some convenient adjustments.
}
\description{
This command is used to fetch budget data using the YNAB api.
}
