% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ynab_get_budgets.R
\name{ynab_get_budgets}
\alias{ynab_get_budgets}
\title{Get Budgets}
\usage{
ynab_get_budgets(pat = Sys.getenv("YNAB_PAT"))
}
\arguments{
\item{pat}{YNAB Personal Access Token}
}
\description{
Access to the Budgets YNAB endpoint
}
\examples{
ynab_get_budgets()
}
