% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ynab_functions.R
\name{ynab_list_budgets}
\alias{ynab_list_budgets}
\title{List the available budgets}
\usage{
ynab_list_budgets()
}
\value{
a data frame of the names and id's of all available 
budgets for download.
}
\description{
This function is to be used after setting the ynab_token object. 
It will fetch and display the available budgets associated with 
the personal access token used.
}
