% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ynabr.R
\name{ynabr}
\alias{ynabr}
\title{Access the You Need A Budget API}
\description{
You Need A Budget (YNAB) is budgeting software and this package provides 
access to the YNAB API.  The likely use case is for users wanting to augment 
the YNAB reports and graphs with custom generated ones.
}
