% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBudgetDetails.R
\name{getBudgetDetails}
\alias{getBudgetDetails}
\title{Get budget details}
\usage{
getBudgetDetails(i)
}
\arguments{
\item{i}{name of endpoint (one of: "accounts", "categories", "months", "payees", "payee_locations", "subcategories", "scheduled_transactions", "transactions")}
}
\description{
Gets the following YNAB data for a pre-specified budget:
 "accounts", "categories", "months", "payees", "payee_locations", 
 "subcategories", "scheduled_transactions", "transactions"
}
\examples{
df_transactions <- getBudgetDetails("transactions")
}
\keyword{getBudgetDetails}
