% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBudgetDetails.R
\name{getBudgetDetails}
\alias{getBudgetDetails}
\title{Gets the following YNAB data for a pre-specified budget:
 "accounts", "categories", "months", "payees", "payee_locations", 
 "subcategories", "scheduled_transactions", "transactions"}
\usage{
getBudgetDetails(i)
}
\arguments{
\item{i}{name of endpoint}
}
\description{
Gets the following YNAB data for a pre-specified budget:
 "accounts", "categories", "months", "payees", "payee_locations", 
 "subcategories", "scheduled_transactions", "transactions"
}
\examples{
df_accounts <- getBudgetDetails("accounts")
df_categories <- getBudgetDetails("categories")
df_subcategories <- getBudgetDetails("subcategories")
df_months <- getBudgetDetails("months")
df_payees <- getBudgetDetails("payees")
df_payee_locations <- getBudgetDetails("payee_locations")
df_scheduled_transactions <- getBudgetDetails("scheduled_transactions")
df_transactions <- getBudgetDetails("transactions")
}
\keyword{getBudgetDetails}
