% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEndpoints.R
\name{getEndpoints}
\alias{getEndpoints}
\title{Retrieve endpoints}
\usage{
getEndpoints()
}
\description{
Returns the YNAB primary endpoints as a data frame
}
\examples{
getEndpoints()
}
\keyword{getEndpoints}
