% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStartingData.R
\name{getStartingData}
\alias{getStartingData}
\title{Gets the following YNAB data for a YNAB subscriber:
 "user", "budgets"}
\usage{
getStartingData(i)
}
\arguments{
\item{i}{name of endpoint}
}
\description{
Gets the following YNAB data for a YNAB subscriber:
 "user", "budgets"
}
\examples{
df_user <- getStartingData("user")
df_budgets <- getStartingData("budgets") 
}
\keyword{getStartingData}
