% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStartingData.R
\name{getStartingData}
\alias{getStartingData}
\title{Retrieves user or budget names}
\usage{
getStartingData(i, param.token)
}
\arguments{
\item{i}{name of endpoint}

\item{param.token}{Your YNAB API personal access token}
}
\description{
Gets the following YNAB data for a YNAB subscriber:
 "user", "budgets"
}
\examples{
endpoint <- "budgets"
mytoken <- "1234567890ABCDE"
df_budgets <- getStartingData(i = endpoint, param.token = mytoken) 
}
\keyword{getStartingData}
