% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYNAB.R
\name{getYNAB}
\alias{getYNAB}
\title{Connect to YNAB API}
\arguments{
\item{param.url}{Character vector of the endpoint URL of interest}

\item{param.token}{Your YNAB API personal access token}
}
\description{
Connects to YNAB API and converts JSON data to R data frames
JSON content > text format > list format > R data frames
Caution: Must be careful with use, as to not overal YNAB's servers
}
\examples{
url <- "https://api.youneedabudget.com/v1/user"
mytoken <- "1234567890ABCDE"
df <- getYNAB(param.url = url, param.token = mytoken)
}
\keyword{getYNAB}
