% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobile-suica.R
\name{parse_mobile_suica_data}
\alias{parse_mobile_suica_data}
\title{Parse Mobile SUICA data}
\usage{
parse_mobile_suica_data(.data, year, purchase_label = "物販")
}
\arguments{
\item{.data}{dataframe of data imported from Mobile SUICA}

\item{year}{year of data}
}
\value{
clean dataframe
}
\description{
Parse Mobile SUICA data
}
