% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refreshTransactions.R
\name{refreshTransactions}
\alias{refreshTransactions}
\title{Updates the YNAB transaction data
Adds new transactions, deletes exisiting or modifies existing transactions.
If there is no df_transactions data frame, it will create one.
Caution: Must be careful with use, as to not overal YNAB's servers}
\usage{
refreshTransactions()
}
\description{
Updates the YNAB transaction data
Adds new transactions, deletes exisiting or modifies existing transactions.
If there is no df_transactions data frame, it will create one.
Caution: Must be careful with use, as to not overal YNAB's servers
}
\examples{
df_transactions <- refreshTransactions()
}
\keyword{refreshTransactions}
