% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectBudget.R
\name{selectBudget}
\alias{selectBudget}
\title{Select a budget}
\usage{
selectBudget()
}
\description{
A function that asks the user to select the budget of interest
}
\examples{
budget_name_id <- selectBudget()
}
\keyword{selectBudget}
