% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ynab_functions.R
\name{ynab_get_account_data}
\alias{ynab_get_account_data}
\title{ynab_get_account_data}
\usage{
ynab_get_account_data(bd)
}
\arguments{
\item{bd}{}
}
\value{

}
\description{
This takes a budget data object as an argument and returns all of the the
transactional account data for all accounts in that budget. Depending on the
amount of data, this can take some time.
}
