% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ynab_functions.R
\name{ynab_set_token}
\alias{ynab_set_token}
\title{Set your YNAB API options}
\usage{
ynab_set_token(token, api_version = "v1")
}
\arguments{
\item{token}{}
}
\value{
No return value
}
\description{
This function sets a global option ynab_token for use in
many other function calls.  It should be the first function
called as other functions will not work without a proper
token. It also sets the base_url option, including the YNAB api version.
}
\details{
The token is obtained through a registration process on the YNAB website. The
token is a 64 character long alphanumeric string.
}
\examples{
## set_ynab_token("test**************************************************")
}
